# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Sistema Inmobiliario",
        "description": "Documentación de la API para gestión de propiedades, inquilinos y contratos.",
        "contact": {
            "email": "admin@test.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/propiedades": {
            "get": {
                "tags": [
                    "Propiedades"
                ],
                "summary": "Listar todas las propiedades",
                "operationId": "1428dc672691f51f4a853ddcf6faff0e",
                "responses": {
                    "200": {
                        "description": "Lista de propiedades recuperada con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "direccion": {
                                                "type": "string",
                                                "example": "Av. Siempre Viva 742"
                                            },
                                            "tipo": {
                                                "type": "string",
                                                "example": "Casa"
                                            },
                                            "estado": {
                                                "type": "string",
                                                "example": "Disponible"
                                            },
                                            "precio_alquiler": {
                                                "type": "number",
                                                "example": 150000
                                            },
                                            "propietario": {
                                                "description": "Datos del dueño",
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Propiedades"
                ],
                "summary": "Crear una nueva propiedad",
                "operationId": "8f14f04fae02e4fe798f67893120cbf3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "propietario_id",
                                    "direccion",
                                    "tipo"
                                ],
                                "properties": {
                                    "propietario_id": {
                                        "description": "ID del dueño existente",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "example": "Av. Siempre Viva 742"
                                    },
                                    "tipo": {
                                        "type": "string",
                                        "example": "Departamento"
                                    },
                                    "ambientes": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "precio_alquiler": {
                                        "description": "Opcional",
                                        "type": "number",
                                        "example": null
                                    },
                                    "comision": {
                                        "description": "Opcional",
                                        "type": "number",
                                        "example": null
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Propiedad creada con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Propiedad creada"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/propiedades/{id}": {
            "get": {
                "tags": [
                    "Propiedades"
                ],
                "summary": "Obtener detalles de una propiedad",
                "operationId": "14c81eabb4a966c2adf109f8a124407e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la propiedad"
                    },
                    "404": {
                        "description": "Propiedad no encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Propiedades"
                ],
                "summary": "Eliminar una propiedad",
                "operationId": "ac46a6152a7d0158329dd1adc227c8b5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la propiedad a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Propiedad eliminada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Propiedad eliminada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Propiedad no encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/propiedades/{id}/fotos": {
            "post": {
                "tags": [
                    "Propiedades"
                ],
                "summary": "Subir una foto a la propiedad",
                "operationId": "7c50385197f4f61f32feb4e1414393fb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la propiedad",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Archivo de imagen",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "imagen": {
                                        "description": "Archivo de imagen (jpg, png, jpeg)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Foto subida con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensaje": {
                                            "type": "string",
                                            "example": "Foto subida"
                                        },
                                        "url": {
                                            "type": "string",
                                            "example": "http://localhost:8000/storage/propiedades/1/foto.jpg"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación (archivo muy grande o formato inválido)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Login con token Bearer",
                "name": "Token de Acceso",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}